created integration files and modified previous methods

+++ src/main/java/com/es2/vadebicicleta/es2_vadebicicleta_equipamento/service/BicicletaService.java
@@ -1,16 +1,22 @@
 package com.es2.vadebicicleta.es2_vadebicicleta_equipamento.service;
 
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.domain.Bicicleta;
+import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.domain.EnderecoEmail;
+import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.domain.Funcionario;
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.domain.StatusBicicletaEnum;
+import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.domain.Totem;
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.domain.Tranca;
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.exception.InvalidActionException;
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.exception.NotFoundException;
+import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.integracao.AluguelClient;
+import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.integracao.ExternoClient;
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.repository.BicicletaRepository;
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.repository.TotemRepository;
 import com.es2.vadebicicleta.es2_vadebicicleta_equipamento.repository.TrancaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 
@@ -21,13 +27,22 @@ public class BicicletaService {
 
     private final TrancaRepository trancaRepository;
 
+    private final TrancaService trancaService;
+
     private final TotemRepository totemRepository;
 
+    private ExternoClient externoClient;
+
+    private AluguelClient aluguelClient;
+
     @Autowired
-    public BicicletaService(BicicletaRepository repository, TrancaRepository trancaRepository, TotemRepository totemRepository) {
+    public BicicletaService(BicicletaRepository repository, TrancaRepository trancaRepository, TrancaService trancaService, TotemRepository totemRepository, ExternoClient externoClient, AluguelClient aluguelClient) {
         this.repository = repository;
         this.trancaRepository = trancaRepository;
+        this.trancaService = trancaService;
         this.totemRepository = totemRepository;
+        this.externoClient = externoClient;
+        this.aluguelClient = aluguelClient;
     }
 
 
@@ -49,6 +64,9 @@ public class BicicletaService {
             return true;
         }   
         if(bicicleta.getStatus() == null || bicicleta.getStatus().isEmpty()){
+            return true;
+        }
+        if(bicicleta.getId() == null && !bicicleta.getStatus().equals("NOVA")){
             return true;
         }
         return bicicleta.getNumero() == null;
@@ -81,6 +99,14 @@ public class BicicletaService {
     }
 
     public void deleteBicicleta(Integer idBicicleta){
+        Bicicleta bicicleta = repository.findById(idBicicleta).orElseThrow(
+            () -> new NotFoundException("Bicicleta não encontrada"));
+        if(!bicicleta.getStatus().equals("APOSENTADA")){
+            throw new NotFoundException("Status Inválido");
+        }
+        if(!trancaRepository.findTrancaByBicicleta(idBicicleta)){
+            throw new NotFoundException("Bicicleta não está associda a nenhuma tranca");
+        }
         if(!repository.deleteById(idBicicleta)){
             throw new NotFoundException("Bicicleta não encontrada");
         }
@@ -120,48 +146,85 @@ public class BicicletaService {
         return bicicleta;
     }
 
+
+    private void enviarEmailInclusao(Funcionario funcionario, Bicicleta bicicleta, Tranca tranca){
+        EnderecoEmail email = new EnderecoEmail();
+        email.setAssunto("Inclusão da bicicleta na rede");
+        email.setMensagem("Data/Hora da Inserção: " + bicicleta.getDataHoraInsercao() + "\nNúmero da Bicicleta: " + bicicleta.getNumero() + "\nNúmero da Tranca: " + tranca.getNumero());
+        email.setEmail(funcionario.getEmail());
+        externoClient.enviarEmail(email);
+    }
+
+    private void enviarEmailRemocao(Funcionario funcionario, Bicicleta bicicleta, Tranca tranca){
+        EnderecoEmail email = new EnderecoEmail();
+        email.setAssunto("Remoção da bicicleta da rede");
+        email.setMensagem("Data/Hora da Remoção: " + bicicleta.getDataHoraInsercao() + "\nNúmero da Bicicleta: " + bicicleta.getNumero() + "\nNúmero da Tranca: " + tranca.getNumero());
+        email.setEmail(funcionario.getEmail());
+        externoClient.enviarEmail(email);
+    }
+
     public void incluirBicicletaNaRedeTotem(Integer idTranca, Integer idBicicleta, Integer idFuncionario){
         Tranca tranca = trancaRepository.findById(idTranca).orElseThrow(
                 () -> new InvalidActionException("Tranca não encontrada"));
-        if(!Objects.equals(tranca.getBicicleta(), idBicicleta)){
-            throw new InvalidActionException("Id da bicicleta na classe Tranca não corresponde ao parâmetro");
+        if(Objects.equals(tranca.getBicicleta(), idBicicleta)){
+            throw new InvalidActionException("Id da bicicleta na classe Tranca corresponde ao id passado como parâmetro");
         }
-        if(idFuncionario == null){
-            throw new InvalidActionException("Funcionário não existe");
+        if(tranca.getBicicleta() != 0){
+            throw new InvalidActionException("Tranca já está associada a uma bicicleta");
         }
+        if(!tranca.getStatus().equals("LIVRE")){
+            throw new InvalidActionException("Status de tranca inválido");
+        }
+        Funcionario funcionario = aluguelClient.obterFuncionario(idFuncionario);
         Integer idTotem = totemRepository.findTotemByTranca(tranca);
         if(idTotem == null){
             throw new InvalidActionException("Tranca não está relacionada com nenhum totem");
         }
-        Bicicleta bicicleta = getById(idBicicleta);
-        if(Objects.equals(bicicleta.getStatus(), "NOVA") || Objects.equals(bicicleta.getStatus(), "DISPONÍVEL")){
-            totemRepository.addTrancasByTotemId(idTotem, tranca);
+        Bicicleta bicicleta = repository.findById(idBicicleta).orElseThrow(
+            () -> new InvalidActionException("Bicicleta não encontrada"));
+        if(Objects.equals(bicicleta.getStatus(), "NOVA")){
+            tranca = trancaService.trancar(idTranca, idBicicleta);
+            bicicleta.setDataHoraInsercao(LocalDateTime.now());
         }
-        else throw new InvalidActionException("Status de bicicleta inválido");
+        if(Objects.equals(bicicleta.getStatus(), "EM_REPARO")){
+            tranca = trancaService.trancar(idTranca, idBicicleta);
+            bicicleta.setDataHoraInsercao(LocalDateTime.now());
+            if(!bicicleta.getFuncionario().equals(idFuncionario)){
+                throw new InvalidActionException("Funcionário não é o mesmo que a colocou em reparo");
+            }
+        }
+        enviarEmailInclusao(funcionario, bicicleta, tranca);
     }
 
     public void retirarBicicletaDaRedeTotem(Integer idTranca, Integer idBicicleta, Integer idFuncionario, String statusAcaoReparador){
         Tranca tranca = trancaRepository.findById(idTranca).orElseThrow(
                 () -> new InvalidActionException("Tranca não encontrada"));
         if(!Objects.equals(tranca.getBicicleta(), idBicicleta)){
-            throw new InvalidActionException("Id da bicicleta na classe Tranca não corresponde ao parâmetro");
+            throw new InvalidActionException("A bicicleta não está associada a uma tranca");
         }
-        if(idFuncionario == null){
-            throw new InvalidActionException("Funcionário não existe");
+        Funcionario funcionario = aluguelClient.obterFuncionario(idFuncionario);
+        if(!tranca.getStatus().equals("OCUPADA")){
+            throw new InvalidActionException("Status da tranca inválido");
         }
         Integer idTotem = totemRepository.findTotemByTranca(tranca);
         if(idTotem == null){
             throw new InvalidActionException("Tranca não está relacionada com nenhum totem");
         }
-        Bicicleta bicicleta = getById(idBicicleta);
-        if(Objects.equals(statusAcaoReparador, "APOSENTADA") || Objects.equals(statusAcaoReparador, "EM_REPARO")){
-            bicicleta.setStatus(statusAcaoReparador);
-            boolean removido = totemRepository.removeTrancaByTotemId(idTotem, tranca);
-            if(!removido){
-                throw new InvalidActionException("Bicicleta não está associada ao totem");
-            }
+        Bicicleta bicicleta = repository.findById(idBicicleta).orElseThrow(
+            () -> new InvalidActionException("Bicicleta não encontrada"));
+        trancaService.destrancar(idTranca, idBicicleta);
+        bicicleta.getDataHoraInsercao();
+        if(!bicicleta.getStatus().equals("REPARO_SOLICITADO")){
+            throw new InvalidActionException("Status da bicicleta inválido");
         }
-        else throw new InvalidActionException("Status de bicicleta inválido");
+        if(Objects.equals(statusAcaoReparador, "APOSENTADA")){
+            bicicleta.setStatus(statusAcaoReparador);
+        }
+        if(Objects.equals(statusAcaoReparador, "EM_REPARO")){
+            bicicleta.setStatus(statusAcaoReparador);
+            bicicleta.setFuncionario(idFuncionario);
+        }
+        enviarEmailRemocao(funcionario, bicicleta, tranca):
     }
 }
